Add final Inventory Controller Remove unnecessary fields for smaller responses sizes

+++ database/migrations/2016_02_17_032800_create_products_table.php
@@ -23,7 +23,7 @@ class CreateProductsTable extends Migration
             $table->integer('breed_id');
             $table->float('price')->nullable();
             $table->integer('quantity')->nullable();
-            $table->integer('adg')->nullable();
+            $table->float('adg')->nullable();
             $table->float('fcr')->nullable();
             $table->float('backfat_thickness')->nullable();
             $table->text('other_details')->nullable();
